re-en(lfs): Large-file support
In CMake Include common project header file with using the modules
search path CMAKE_MODULE_PATH instead of include the header file direct
with its relative path.

Simplify handling of LFS and use C++ features for it.

Document LFS on the overview page.

Modify github Actions job to compile 64 bit architectures.

+++ Compa/src/genlib/net/http/httpreadwrite.cpp
@@ -4,7 +4,7 @@
  * All rights reserved.
  * Copyright (c) 2012 France Telecom All rights reserved.
  * Copyright (C) 2022+ GPL 3 and higher by Ingo Höft, <Ingo@Hoeft-online.de>
- * Redistribution only with this Copyright remark. Last modified: 2024-08-18
+ * Redistribution only with this Copyright remark. Last modified: 2024-09-07
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -65,6 +65,8 @@
 #else /* _WIN32 */
 /// \cond
 #include <sys/utsname.h>
+// fseeko is not supported on 32-bit Android older than API 24:
+// https://android.googlesource.com/platform/bionic/+/main/docs/32-bit-abi.md
 #if defined(__ANDROID__) &&                                                    \
     (!defined(__USE_FILE_OFFSET64) || __ANDROID_API__ < 24)
 #define fseeko fseek
@@ -856,7 +858,7 @@ int http_SendMessage(SOCKINFO* info, int* TimeOut, const char* fmt, ...) {
                     /* Hex length for the chunk size. */
                     memset(Chunk_Header, 0, sizeof(Chunk_Header));
                     rc = snprintf(Chunk_Header, sizeof(Chunk_Header),
-                                  "%" PRIzx "\r\n", (unsigned long)num_read);
+                                  "%" PRIzx "\r\n", num_read);
                     if (rc < 0 || (unsigned int)rc >= sizeof(Chunk_Header)) {
                         RetVal = UPNP_E_INTERNAL_ERROR;
                         goto Cleanup_File;
@@ -1564,8 +1566,7 @@ int http_MakeMessage(membuffer* buf, int http_major_version,
         } else if (c == 'd') {
             /* integer */
             num = (size_t)va_arg(argp, int);
-            rc = snprintf(tempbuf, sizeof(tempbuf), "%" PRIzu,
-                          (unsigned long)num);
+            rc = snprintf(tempbuf, sizeof(tempbuf), "%" PRIzu, num);
             if (rc < 0 || (unsigned int)rc >= sizeof(tempbuf) ||
                 membuffer_append(buf, tempbuf, strlen(tempbuf)))
                 goto error_handler;
